requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==0.21.1
sqlalchemy
requests